Finished refactoring WFSRequest to use the same shared class. It may be possible to further simplify the functions to use the base functionality
