+++ Assets/_Functionalities/OGCWebServices/Wms/Scripts/GetMapRequest.cs
@@ -10,6 +10,8 @@ namespace Netherlands3D.Functionalities.Wms
 {
     public class GetMapRequest : BaseRequest
     {
+        private const string defaultFallbackVersion = "1.3.0"; //todo: make shared between wms classes?
+     
         protected override Dictionary<string, string> defaultNameSpaces => OgcWebServicesUtility.DefaultWmsNamespaces;
         
         public GetMapRequest(Uri sourceUrl, string xml) : base(sourceUrl, xml)
